requirements.txt
aiofiles==23.1.0
anyio==3.6.2
asgiref==3.6.0
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Django==4.1.7
fastapi==0.95.0
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
paramiko==3.1.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.7
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
sshtunnel==0.4.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.2
uvicorn==0.21.1
